Secure API endpoint Fixes #1
Refactor `schema.py` into schema module with base and file for each django model
