+++ grocerfinder_root/aldi/schema/__init__.py
@@ -1,0 +1,9 @@
+from . import item, category
+
+
+class Query(item.Query, category.Query):
+    pass
+
+
+class Mutation(item.Mutation, category.Mutation):
+    pass
